services/telemetry/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2 